Update folder structure
