--- library/index.js
@@ -1,6 +1,0 @@
-//
-// Copyright (c) 2022-2023 Ivan Teplov
-// Licensed under the Apache license 2.0
-//
-
-export * from "./sheet.jsx"
